+++ src/main/java/net/minecraftforge/event/ForgeEventFactory.java
@@ -125,7 +125,6 @@ import net.minecraftforge.event.world.BlockEvent.NeighborNotifyEvent;
 import net.minecraftforge.event.world.BlockEvent.PlaceEvent;
 import net.minecraftforge.event.world.ExplosionEvent;
 import net.minecraftforge.event.world.FoundChunksForSpawningEvent;
-import net.minecraftforge.event.world.FoundChunksForSpawningEventBuilder;
 import net.minecraftforge.event.world.GetCollisionBoxesEvent;
 import net.minecraftforge.event.world.WorldEvent;
 import net.minecraftforge.fml.common.ObfuscationReflectionHelper;
@@ -308,7 +307,7 @@ public class ForgeEventFactory
         return maxCanSpawnEvent.getResult() == Result.ALLOW ? maxCanSpawnEvent.getMaxPackSize() : entity.getMaxSpawnedInChunk();
     }
 
-    public static void onFoundChunksForSpawning(FoundChunksForSpawningEventBuilder builder)
+    public static void onFoundChunksForSpawning(FoundChunksForSpawningEvent.Builder builder)
     {
         MinecraftForge.EVENT_BUS.post(new FoundChunksForSpawningEvent(builder));
     }
